# backend/grip/federation/contract/opdrachtverkeer/v1/openapi.yaml
openapi: 3.1.0
info:
  title: grip-opdrachtverkeer
  version: 1.0.0
  summary: Het verkeer tussen grip-instanties over aanvragen, offertes, akkoord, inzage, vacatures en doorgifte.
  description: |
    Elke grip-instantie biedt deze dienst aan achter de eigen FSC-inway, onder de dienstnaam
    `grip-opdrachtverkeer`. Een instantie kan opdrachtgever en opdrachtnemer zijn. Een opdracht is
    lokaal volledig; dit verkeer is een laag erbovenop.

    **Identiteit.** Opdrachten, offertes en vacatures hebben een URI van de instantie die ze
    aanmaakt, bijvoorbeeld `{basis}/id/opdracht/{uuid}`. De `{opdrachtId}`, `{offerteId}` en
    `{vacatureId}` in de paden zijn de `{uuid}` uit die URI. Organisaties worden aangeduid met hun
    TOOI-URI.

    **Aanroeper.** De aanroeper is een FSC-peer. De ontvanger leidt de relatie af uit het peer-id:
    opdrachtgever of opdrachtnemer van die opdracht, moederinstantie, of corpus. Elke operatie
    noemt in `x-grip-caller` welke relatie vereist is. Een opdracht waar de aanroeper geen relatie
    mee heeft geeft 404.

    **Gegevensklassen.** Elke operatie noemt in `x-grip-data-class` de klassen die het bericht of
    antwoord bevat: A opdracht-basis, B opdracht-financieel, C bemensing, D persoon-tarief,
    E persoon-kosten, F persoon-KPI. De klassen D, E en F gaan nooit over de grens van een
    instantie. Naar een opdrachtgever gaan geen namen van medewerkers; klasse C met namen gaat
    alleen naar de moederinstantie.

...

...

...

...
  contact:
...
servers:
...
paths:
...
  /opdrachten/{opdrachtId}/voortgang:
    get:
      operationId: getProgress
      tags: [Inzage]
      summary: Status en voortgang
      description: Standaard opvraagbaar door de opdrachtgever. Geen bedragen en geen namen.
      x-grip-caller: client
      x-grip-data-class: [A]
      parameters:
        - { $ref: "#/components/parameters/opdrachtId" }
      responses:
        "200":
          description: De voortgang.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/voortgang.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /opdrachten/{opdrachtId}/uitputting:
    get:
      operationId: getBudgetUsage
      tags: [Inzage]
      summary: Uitputting
      description: Begroot, uitputting en beschikbaar. Alleen op verzoek; 403 als de opdrachtnemer dit voor deze opdrachtgever niet heeft toegestaan.
      x-grip-caller: client
      x-grip-data-class: [B]
      x-grip-on-request: true
      parameters:
        - { $ref: "#/components/parameters/opdrachtId" }
        - name: jaar
          in: query
          description: Kalenderjaar. Zonder deze parameter de hele looptijd.
          schema: { type: integer, minimum: 2000 }
      responses:
        "200":
          description: De uitputting.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/uitputting.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /jwks:
    get:
      operationId: getJwks
      tags: [Sleutels]
      summary: Publieke sleutels van deze instantie
      description: De sleutels waarmee deze instantie akkoordberichten ondertekent. Een ontvanger haalt ze op bij de afzender om een JWS te controleren. Een sleutel die niet meer wordt gebruikt blijft opvraagbaar zolang er akkoorden mee zijn getekend.
      x-grip-caller: peer
      x-grip-data-class: []
      responses:
        "200":
          description: De sleutelset.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/jwks.schema.json" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  parameters:
    opdrachtId:
      name: opdrachtId
      in: path
      required: true
      description: De uuid uit de opdracht-URI.
      schema: { type: string, format: uuid }
...
  headers:
    API-Version:
      description: Volledige versie van het contract dat deze implementatie volgt.
      schema: { type: string, examples: ["1.0.0"] }
  responses:
...
    BadRequest:
      description: Het bericht of verzoek is ongeldig, bijvoorbeeld omdat het niet aan het schema voldoet of de hash niet klopt met de inhoud.
      content:
        application/problem+json:
          schema: { $ref: "../../schemas/problem.schema.json" }
    Forbidden:
      description: De aanroepende peer heeft niet de vereiste relatie, of de gevraagde inzage is niet toegestaan.
      content:
        application/problem+json:
          schema: { $ref: "../../schemas/problem.schema.json" }
    NotFound:
      description: Niet gevonden, of de aanroeper heeft er geen relatie mee.
      content:
        application/problem+json:
          schema: { $ref: "../../schemas/problem.schema.json" }
...